CodeGen: Improve naming of copied basic blocks
It may happen that we generate the code of a basic block from the original
scop is code generated several times. The new naming scheme reduces confusing
that earlier appeared as the version numbers of the new basic blocks could
have been interpreted as part of the name of the original basic block.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@139092 91177308-0d34-0410-b5e6-96231b3b80d8