RegisterPasses: Do not claim we schedule canonicalization passes at -O0
Also the code makes the impression this was happening, shouldEnablePolly()
always returns false for optlevel equal to zero. This was previously different,
but was accidentally changed by a commit a couple of months ago. As this
behavior was mainly a debugging tool and adding this to clang never really made
sense, we just remove the last traces.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@198370 91177308-0d34-0410-b5e6-96231b3b80d8