Do not track location of scalar dependences in ScopInfo
We do not have a use for this information at the moment. If we need this at some
point, the "instruction -> access" mapping needs to be enhanced as a single
instruction could then possibly perform multiple accesses.

This patch allows us to build the polyhedral information for scops with scalar
dependences.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@201815 91177308-0d34-0410-b5e6-96231b3b80d8