ScopDetection: Support keep-going
Support a 'keep-going' mode for the ScopDetection. In this mode, we just keep
on detecting, even if we encounter an error.

This is useful for diagnosing SCoP candidates. Sometimes you want all the
errors. Invalid SCoPs will still be refused in the end, we just refuse to
abort on the first error.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@209574 91177308-0d34-0410-b5e6-96231b3b80d8