Check for an empty error log.
Fixes #19976.

The error log does not contain an error, in case we reject a candidate
without generating a diagnostic message by using invalid<>(...). This is
the case for the top-level region of a function.

The patch comes without a test-case because adding a useful one requires
additional code just for triggering it. Before the patch it would only trigger,
if we try to print the CFG with Scop error annotations.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@210753 91177308-0d34-0410-b5e6-96231b3b80d8