Always pipe in test files
In Polly we used to have a mix of test cases, some that used 'opt %s' and others
that used 'opt < %s'. We now change all to use 'opt < %s'. Piping in test files
is preferable as it does prevent temporary files to be written to disk. This
brings us in line with what is usus in LLVM.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@216816 91177308-0d34-0410-b5e6-96231b3b80d8