Report possible aliasing deterministically
This commit drops a call to std::sort, which sorted the base pointers that
possibly alias according to the address at which their corresponding llvm::Value
was allocated. There does not seem to be any good reason, why those pointers
should be (re)sorted and this only makes the output indeterministic.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@218052 91177308-0d34-0410-b5e6-96231b3b80d8