Delete some unnecessary code
Originally we have needed this code to map the isl_id of an array to its base
pointer. However, as now the isl_id contains a reference to the array itself we
obtain the base pointer from this isl_id and we do not need to add this
information to the IDToValue map.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@220876 91177308-0d34-0410-b5e6-96231b3b80d8