Use explicit StringRef Type
We previously used a Twine here, but as pointed out by David Blaikie
and Mehdi Amini storing a temporary StringRef in a Twine is not a good
idea, as the StringRef will be freed before the Twine is used leaving
a Twine that points to uninitialized memory. We now make it explicit that
we use a StringRef here.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@225342 91177308-0d34-0410-b5e6-96231b3b80d8