Pass an actual schedule to the isl_union_access [NFC]
This change is a step towards using a single isl_schedule object throughout
Polly. At the moment the schedule is just constructed from the flat
isl_union_map that defines the schedule. Later we will obtain it directly
from the scop and potentially obtain a schedule with a non-trivial internal
structure that will allow faster dependence analysis.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@235378 91177308-0d34-0410-b5e6-96231b3b80d8