Replace repository version of ISL by 'make dist' output
The 'make dist' archive is not dependent on ./configure output and
contains a GIT_HEAD_ID file that identifies the version of ISL used.

None of the files added or removed are used part of Polly's build
process (except of GIT_HEAD_ID since the previous revision r240301). No
functional change intended.



git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@240306 91177308-0d34-0410-b5e6-96231b3b80d8