Replace srem by function call in nonaffine test cases
This makes the test cases nonaffine even if Polly some days gains support for
the srem instruction, an instruction which is currently not modeled but which
can clearly be modeled statically. A call to a function without definition
will always remain non-affine, as there is just insufficient static information
for it to be modeled more precisely.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@240458 91177308-0d34-0410-b5e6-96231b3b80d8