Fix Codegen adding a second exit out of region
executeScopConditionally would destroy a predecessor region if it the
scop's entry was the region's exit block by forking it to polly.start
and thus creating a secnd exit out of the region. This patch "shrinks"
the predecessor region s.t. polly.split_new_and_old is not the 
region's exit anymore. 



git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@245294 91177308-0d34-0410-b5e6-96231b3b80d8