Add support for two-level tiling
By default we only use one level of tiling for loops, but in general tiling
for multiple levels is trivial for us. Hence, we add a set of options that
allow people to play with a second level of tiling. If this is profitable for
some cases we can work on heuristics that allow us to identify these cases
and use two-level tiling for them.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@245563 91177308-0d34-0410-b5e6-96231b3b80d8