Do not model delinearized and linearized access relation for a single access
A missing return statement that previously did not have a visibly negative
effect caused after some data-structure changes in r248024 multi-dimensional
accesses to be modeled both multi-dimensional as well as linearized. This
commit adds the missing return to avoid the incorrect double modeling as
well as the compile time increases it caused.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@248171 91177308-0d34-0410-b5e6-96231b3b80d8