IRBuilder: Ensure we do not add empty map elements
Do not use "Map[Key] == nullptr" to check if a Key is in the map, but use
"Map.find(Key) == Map.end()". Map[Key] always adds Key into the map, a
side-effect we do not want.

Found by inspection. This is hard to test outside of a targetted unit test,
which seems too much overhead for this individual issue.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@249544 91177308-0d34-0410-b5e6-96231b3b80d8