Fix escaping value to subregion entry node phi
An incoming value from a block the is not inside the scop is an
external use, even if the phi is inside the scop. A previous fix in
r251208 did not apply if the phi is inside a non-affine subregion. We
move the check for this phi case before the non-affine subregion check.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@252157 91177308-0d34-0410-b5e6-96231b3b80d8