ScopInfo: Look up first (and only) array access
When rewriting the access functions of load/store statements, we are only
interested in the actual array memory location. The current code just took
the very first memory access, which could be a scalar or an array access. As
a result, we failed to update access functions even though this was requested
via .jscop.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@255713 91177308-0d34-0410-b5e6-96231b3b80d8