Add option to disallow modref function calls in scops.
This might be useful to evaluate the benefit of us handling modref funciton
calls. Also, a new bug that was triggered by modref function calls was
recently reported http://llvm.org/PR27035. To ensure the same issue does not
cause troubles for other people, we temporarily disable this until the bug
is resolved.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@264140 91177308-0d34-0410-b5e6-96231b3b80d8