Recommit: "Simplify min/max expression generation"
As part of this simplification we pull complex logic out of the loop body and
skip the previously redundantly executed first loop iteration.

This is a partial recommit of r271514 and r271535 which where in conflict with
the revert in r272483 and consequently also had to be reverted temporarily.  The
original patch was contributed by Johannes Doerfert.

This patch is mostly a NFC, but dropping the first loop iteration can sometimes
result in slightly simpler code.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@272502 91177308-0d34-0410-b5e6-96231b3b80d8