GPGPU: Use schedule whole components for scheduler
This option increases the scalability of the scheduler and allows us to remove
the 'gisting' workaround we introduced in r275565 to handle a more complicated
test case. Another benefit of using this option is also that the generated
code looks a lot more streamlined.

Thanks to Sven Verdoolaege for reminding me of this option.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@275573 91177308-0d34-0410-b5e6-96231b3b80d8