GPURuntime: Drop polly_cleanupGPGPUResources
This function is currently unused and won't be used in this form again. Instead
of freeing many unrelated items at the same time, we will instead explicitly
call free function from the host-IR we generate for each object we want to free.
These specific free functions will be added together with the corresponding
host-IR generation code.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@276632 91177308-0d34-0410-b5e6-96231b3b80d8