[GPGPU] Use separate basic block for GPU initialization code
This increases the readability of the IR and also clarifies that the GPU
inititialization is executed _after_ the scalar initialization which needs
to before the code of the transformed scop is executed.

Besides increased readability, the IR should not change. Specifically, I
do not expect any changes in program semantics due to this patch.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@278125 91177308-0d34-0410-b5e6-96231b3b80d8