Introduce unittests.
Add the infrastructure for unittests to Polly and two simple tests for
conversion between isl_val and APInt. In addition, a build target
check-polly-unittests is added to run only the unittests but not the regression
tests.

Clang's unittest mechanism served as as a blueprint which then was adapted to
Polly.

Differential Revision: https://reviews.llvm.org/D23833

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@279734 91177308-0d34-0410-b5e6-96231b3b80d8