test: add more details to non-affine test case
We add CHECK lines to this test case to make it easier to see the difference
between affine and non-affine memory accesses. We also change the test case to
use a parameteric index expression as otherwise our range analysis will
understand that the non-affine memory access can only access input[1],
which makes it difficult to see that the memory access is in-fact modeled as
non-affine access.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@287623 91177308-0d34-0410-b5e6-96231b3b80d8