[ScopInfo] Do not try to fold array dimensions of size zero
Trying to fold such kind of dimensions will result in a division by zero,
which crashes the compiler. As such arrays are likely to invalidate the
scop anyhow (but are not illegal in LLVM-IR), there is no point in trying
to optimize the array layout. Hence, we just avoid the folding of
constant dimensions of size zero.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@295415 91177308-0d34-0410-b5e6-96231b3b80d8