[ScopInfo] Count read-only arrays when computing complexity of alias check
Instead of counting the number of read-only accesses, we now count the number of
distinct read-only array references when checking if a run-time alias check
may be too complex. The run-time alias check is quadratic in the number of
base pointers, not the number of accesses.

Before this change we accidentally skipped SPEC's lbm test case.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@295567 91177308-0d34-0410-b5e6-96231b3b80d8