Remove all references to PostDominators. NFC.
Marking a pass as preserved is necessary if any Polly pass uses it, even
if it is not preserved within the generated code. Not marking it would
cause the the Polly pass chain to be interrupted. It is not used by any
Polly pass anymore, hence we can remove all references to it.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@295983 91177308-0d34-0410-b5e6-96231b3b80d8