[unittests/DeLICM] Add test for Occipied vs Occupied.
The interpretation of multiple known ValInsts for the same element and
timepoint is that these are alterntivate names for the same values,
for instance a PHINode and the incoming value when knowning it was
the last executed block. That means that known values do not conflict
if there at least (but necessarily all) one common ValInst.

Add a case to test this principle.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@301480 91177308-0d34-0410-b5e6-96231b3b80d8