[ScopDetection] Check for already known required-invariant loads [NFC]
For certain test cases we spent over 50% of the scop detection time in
checking if a load is likely invariant. We can avoid most of these checks by
testing early on if a load is expected to be invariant. Doing this reduces
scop-detection time on a large benchmark from 52 seconds to just 25 seconds.

No functional change is expected.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@302134 91177308-0d34-0410-b5e6-96231b3b80d8