[GPUJIT] Disabled gcc's -Wpedantic for use of dlsym
GCC's ISO C standard does not strictly define the bahavior of converting
a `void*` pointer to a function pointer, but dlsym's POSIX standard
does.

The retrieval of function pointers through dlsym in this case
generates an unnecessary amount of warnings for every API function
assignment, bloating the output.

This patch removes GCC's `-Wpedantic` flag for retrieval and assignment
of these functions. This simplifies debugging the output of GPUJIT.

Differential Revision: https://reviews.llvm.org/D33008

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@302638 91177308-0d34-0410-b5e6-96231b3b80d8