[Simplify] Use execution order of memory accesses.
Iterate through memory accesses in execution order (first all implicit reads,
then explicit accesses, then implicit writes).

In the test case this caused an implicit load to be handled as if it was loaded
after the write. That is, the value being written before it is available.

This fixes llvm.org/PR33323

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@304810 91177308-0d34-0410-b5e6-96231b3b80d8