[ScopDetect/Info] Look through PHIs that follow an error block
In case a PHI node follows an error block we can assume that the incoming value
can only come from the node that is not an error block. As a result, conditions
that seemed non-affine before are now in fact affine.

This is a recommit of r312663 after fixing
test/Isl/CodeGen/phi_after_error_block_outside_of_scop.ll

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@314075 91177308-0d34-0410-b5e6-96231b3b80d8