Apply include-what-you-use #include removal suggestions. NFC.
This removes unused includes (and forward declarations) as
suggested by include-what-you-use. If a transitive include of a removed
include is required to compile a file, I added the required header (or
forward declaration if suggested by include-what-you-use).

This should reduce compilation time and reduce the number of iterative
recompilations when a header was changed.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@357209 91177308-0d34-0410-b5e6-96231b3b80d8