[DeLICM] Use polly::singleton to allow empty result.
isl_map_from_union_map cannot determine the map's space if the union_map
is empty. polly::singleton was designed for this case. We pass the
expected map space to avoid crashing in isl_map_from_union_map.

This fixes an issue found by the aosp buildbot. Thanks to Eli Friedman
for the reproducer.

git-svn-id: https://llvm.org/svn/llvm-project/polly/trunk@361290 91177308-0d34-0410-b5e6-96231b3b80d8