Merging r372188:
------------------------------------------------------------------------
r372188 | meinersbur | 2019-09-17 15:59:43 -0700 (Tue, 17 Sep 2019) | 13 lines

[CodeGen] Handle outlining of CopyStmts.

Since the removal of extensions nodes from schedule trees in r362257 it
is possible to emit parallel code for SCoPs containing
matrix-multiplications. However, the code looking for references used in
outlined statement was not prepared to handle CopyStmts introduced by
the matrix-matrix multiplication detection.

In this case, CopyStmts do not introduce references in addition to the
ones captured by MemoryAccesses, i.e. we change the assertion to accept
CopyStmts and add a regression test for this case.

This fixes llvm.org/PR43164
------------------------------------------------------------------------

git-svn-id: https://llvm.org/svn/llvm-project/polly/branches/release_90@374861 91177308-0d34-0410-b5e6-96231b3b80d8